% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_plot_report_generic.R
\name{create_code_chunk}
\alias{create_code_chunk}
\title{Generate R-code Chunk to include Plot Pairs}
\usage{
create_code_chunk(ps_plot_file, ps_left_plot_path, ps_right_plot_path)
}
\arguments{
\item{ps_plot_file}{name of the plot file}

\item{ps_left_plot_path}{path to plot on the left-hand-side}

\item{ps_right_plot_path}{path to the plot on the right-hand-side}
}
\value{
s_code_chunk_result R-code chunk  as a string which includes the plots
}
\description{
Given the name of the plot file and the paths to the plot pair to be included
in the report, the R-code chunk used in the RMarkdown report is generated and
returned as a string.
}
