% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_plot_report_birth.R
\name{create_ge_compare_plot_report_birth}
\alias{create_ge_compare_plot_report_birth}
\title{Comparison Plot Report Creator Function For Production (BIRTH)}
\usage{
create_ge_compare_plot_report_birth(
  pn_cur_ge_label,
  pn_prev_ge_label,
  ps_cur_plot_root = NULL,
  ps_prev_plot_root = NULL,
  ps_template = system.file("templates", "compare_plots.Rmd.template", package =
    "qgert"),
  pl_plot_opts = NULL,
  pb_debug = FALSE,
  plogger = NULL
)
}
\arguments{
\item{pn_cur_ge_label}{label of current genetic evaluation (GE)}

\item{pn_prev_ge_label}{label of previous GE}

\item{ps_cur_plot_root}{directory with plots of current evaluation}

\item{ps_prev_plot_root}{directory with plots from previous evaluation}

\item{ps_template}{template document for report}

\item{pl_plot_opts}{list of options specifying input for plot report creator}

\item{pb_debug}{flag whether debug output should be shown}

\item{plogger}{log4r logger object}

\item{ps_breed}{create comparison plot report for just one breed}
}
\description{
A comparison plot report containing all generated plots of a GE side-by-side
with the plots from the previous GE are constructed for the trait
group BIRTH
}
\examples{
\dontrun{
create_ge_compare_plot_report_birth(pn_cur_ge_label  = 2008,
                                 pn_prev_ge_label = 2004,
                                 pb_debug = TRUE)
}

}
