% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_plot_report_cnvrh.R
\name{create_ge_compare_plot_report_cnvrh}
\alias{create_ge_compare_plot_report_cnvrh}
\title{Comparison Plot Report Creator Function For Convert RH (CNVRH)}
\usage{
create_ge_compare_plot_report_cnvrh(
  pn_cur_ge_label,
  pn_prev_ge_label,
  ps_template = system.file("templates", "compare_plots.Rmd.template", package =
    "qgert"),
  pl_plot_opts = NULL,
  pb_debug = FALSE,
  plogger = NULL
)
}
\arguments{
\item{pn_cur_ge_label}{label of current genetic evaluation (GE)}

\item{pn_prev_ge_label}{label of previous GE}

\item{ps_template}{template document for report}

\item{pl_plot_opts}{list of options specifying input for plot report creator}

\item{pb_debug}{flag whether debug output should be shown}

\item{plogger}{log4r logger object}
}
\description{
A comparison plot report containing all generated plots of a GE side-by-side
with the plots from the previous GE are constructed for the plots produced
in the conversion of cow proofs for RH (CNVRH).
}
\examples{
\dontrun{
create_ge_compare_plot_report_cnvrh(pn_cur_ge_label  = 1908,
                                 pn_prev_ge_label = 1904,
                                 pb_debug = TRUE)
}

}
