% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_plot_report_generic.R,
%   R/compose_compare_plot.R
\name{create_ge_plot_report}
\alias{create_ge_plot_report}
\title{Create Report With Plots From Two GE Periods}
\usage{
create_ge_plot_report(ps_gedir, ps_archdir, ps_trgdir, ps_templ,
  ps_report_text, ps_rmd_report = "ge_plot_report.Rmd",
  pb_debug = FALSE)

create_ge_plot_report(ps_gedir, ps_archdir, ps_trgdir, ps_templ,
  ps_report_text, ps_rmd_report = "ge_plot_report.Rmd",
  pb_debug = FALSE)
}
\arguments{
\item{ps_gedir}{directory with plots of current GE round}

\item{ps_archdir}{archive directory with plots from previous GE}

\item{ps_templ}{path to Rmarkdown template file}

\item{ps_report_text}{text that is included in the report before plotting}

\item{ps_rmd_report}{name of report source file}

\item{pb_debug}{flag indicating whether debug info is printed}

\item{pstrgdir}{target directory where plot files from archive are extracted, relative to ps_gedir}

\item{plogger}{log4r logger object}

\item{ps_gedir}{directory with plots of current GE round}

\item{ps_archdir}{archive directory with plots from previous GE}

\item{pstrgdir}{target directory where plot files from archive are extracted, relative to ps_gedir}

\item{ps_templ}{path to Rmarkdown template file}

\item{ps_report_text}{text that is included in the report before plotting}

\item{ps_rmd_report}{name of report source file}

\item{pb_debug}{flag indicating whether debug info is printed}
}
\description{
Based on a Rmarkdown Template (ps_templ) document all plots
in given GE directory (ps_gedir) are taken. For a given plot
in directory ps_gedir, a corresponding plot with the same
filename is searched in an archive directory (ps_archdir).
If such a plot is found, the two corresponding plots are
shown side-by-side in the generated Rmarkdown report.

Based on a Rmarkdown Template (ps_templ) document all plots
in given GE directory (ps_gedir) are taken. For a given plot
in directory ps_gedir, a corresponding plot with the same
filename is searched in an archive directory (ps_archdir).
If such a plot is found, the two corresponding plots are
shown side-by-side in the generated Rmarkdown report.
}
\examples{
\dontrun{
create_ge_plot_report(ps_gedir       = "/Volumes/data_zws/fbk/work/bv/YearMinus0/compareBull",
                    ps_archdir     = "/Volumes/data_archiv/zws/1904/fbk/work/bv/YearMinus0/compareBull",
                    ps_trgdir      = "1904/compareBull",
                    ps_templ       = "inst/templates/compare_plots.Rmd.template",
                    ps_report_text = "## Comparison Of Plots\\nPlots compare estimates of fbk for bulls of breed BV between GE-1904 on the left and the current GE-1908 on the right.",
                    ps_rmd_report  = 'ge_plot_report_fbk_compareBulls_bv.Rmd',
                    pb_debug       = TRUE)

}
\dontrun{
create_ge_plot_report(ps_gedir       = "/Volumes/data_zws/fbk/work/bv/YearMinus0/compareBull",
                    ps_archdir     = "/Volumes/data_archiv/zws/1904/fbk/work/bv/YearMinus0/compareBull",
                    ps_trgdir      = "1904/compareBull",
                    ps_templ       = "inst/templates/compare_plots.Rmd.template",
                    ps_report_text = "## Comparison Of Plots\\nPlots compare estimates of fbk for bulls of breed BV between GE-1904 on the left and the current GE-1908 on the right.",
                    ps_rmd_report  = 'ge_plot_report_fbk_compareBulls_bv.Rmd',
                    pb_debug       = TRUE)

}
}
