% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toplist_bull_bvch.R
\name{create_toplist_bvch_bull}
\alias{create_toplist_bvch_bull}
\title{Create Toplists for BVCH-Bulls}
\usage{
create_toplist_bvch_bull(
  ps_eval_label,
  pl_breed_input = get_l_breed_input_bvch(),
  ptbl_trait_name = get_default_tbl_trait_name_bvch(),
  ps_xlsx_file = NULL,
  pb_debug = FALSE,
  plogger = NULL
)
}
\arguments{
\item{ps_eval_label}{evaluation label}

\item{pl_breed_input}{list with input items consisting of breed names, input files and length of rankings}

\item{pb_debug}{flag indicating debugging mode}

\item{plogger}{logger object}
}
\description{
Based on information for the breeds consisting of input files, breed names and
number of records that should be listed in the toplist and a tibble with trait
names, the results are extracted from the input files and are then written to
an xlsx-file.
}
\details{
---
title: Create Toplists for BVCH-Bulls
date: 2020-08-11
---
}
