% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_plot_report_constant.R
\name{get_default_plot_opts_birth}
\alias{get_default_plot_opts_birth}
\title{Default Plot Options For BIRTH}
\usage{
get_default_plot_opts_birth()
}
\description{
Return a list with specific defaults and constants that are used
to produce the comparison plot report for the trait group
BIRTH
}
