% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_plot_report_constant.R
\name{get_default_plot_opts_calving_beef}
\alias{get_default_plot_opts_calving_beef}
\title{Default Plot Options For CALVING BEEF}
\usage{
get_default_plot_opts_calving_beef()
}
\description{
Return a list with specific defaults and constants that are used
to produce the comparison plot report for the trait group
CALVING BEEF
}
