% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_plot_report_constant.R
\name{get_default_plot_opts_carcass}
\alias{get_default_plot_opts_carcass}
\title{Default Plot Options For CARCASS}
\usage{
get_default_plot_opts_carcass()
}
\description{
Return a list with specific defaults and constants that are used
to produce the comparison plot report for the trait group
CARCASS
}
