% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_plot_report_constant.R
\name{get_default_plot_opts_cas}
\alias{get_default_plot_opts_cas}
\title{Default Plot Options For CAS}
\usage{
get_default_plot_opts_cas()
}
\description{
Return a list with specific defaults and constants that are used
to produce the plot comparison report for the de/regression (CAS).
}
