% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_plot_report_constant.R
\name{get_default_plot_opts_fbk}
\alias{get_default_plot_opts_fbk}
\title{Default Plot Options For FBK}
\usage{
get_default_plot_opts_fbk()
}
\description{
Return a list with specific defaults and constants that are used
to produce the plot comparison report for the trait group
Fruchtbarkeit (FBK).
}
