% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_plot_report_constant.R
\name{get_default_plot_opts_gs_dereg}
\alias{get_default_plot_opts_gs_dereg}
\title{Default Plot Options For GS DEREG}
\usage{
get_default_plot_opts_gs_dereg()
}
\description{
Return a list with specific defaults and constants that are used
to produce the plot comparison report for the de/regression (GS DEREG).
}
