% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_plot_report_constant.R
\name{get_default_plot_opts_ker}
\alias{get_default_plot_opts_ker}
\title{Default Plot Options For KER}
\usage{
get_default_plot_opts_ker()
}
\description{
Return a list with specific defaults and constants that are used
to produce the plot comparison report for the de/regression (KER).
}
