% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_plot_report_constant.R
\name{get_default_plot_opts_lbe_rh}
\alias{get_default_plot_opts_lbe_rh}
\title{Default Plot Options For LBE_RH}
\usage{
get_default_plot_opts_lbe_rh()
}
\description{
Return a list with specific defaults and constants that are used
to produce the comparison plot report for the trait group
Lineare Beschreibung for RH (LBE_RH).
}
