% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_plot_report_constant.R
\name{get_default_plot_opts_prod}
\alias{get_default_plot_opts_prod}
\title{Default Plot Options For PROD}
\usage{
get_default_plot_opts_prod()
}
\description{
Return a list with specific defaults and constants that are used
to produce the comparison plot report for the trait group
Production (PROD).
}
