% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_plot_report_constant.R
\name{get_default_plot_opts_vrdggozw_prov}
\alias{get_default_plot_opts_vrdggozw_prov}
\title{Default Plot Options For VRDGGOZW_PROV}
\usage{
get_default_plot_opts_vrdggozw_prov()
}
\description{
Return a list with specific defaults and constants that are used
to produce the comparison plot report for preliminary genomic breeding values
(VRDGGOZW_PROV) based on previous run ITB_BVs.
}
