% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toplist_bull_bvch.R
\name{get_default_tbl_trait_name_bvch}
\alias{get_default_tbl_trait_name_bvch}
\title{Default Trait Information}
\usage{
get_default_tbl_trait_name_bvch()
}
\value{
Tibble with default information on traits for bvch
}
\description{
The default version of trait information is read from a csv-file in the package.
If different traits must be considered, then they have to be specified explicitly
in the top creation function.
}
\details{
--- Determine Trait Names -------------------------------------------------- #
}
