% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toplist_bull_bvch.R
\name{get_l_breed_input_bvch}
\alias{get_l_breed_input_bvch}
\title{Breed Input for Breed BVCH}
\usage{
get_l_breed_input_bvch()
}
\value{
List of information about breeds
}
\description{
This function tries to return reasonable default values for the information
about the breeds.
}
\details{
If nothing is specified, the toplist input is taken from the extdata directory
of the package. Hence this is only to be used for testing.
}
