% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_plot_report_generic.R
\name{get_plot_files}
\alias{get_plot_files}
\title{Getting a List of Plot Files}
\usage{
get_plot_files(ps_plot_dir)
}
\arguments{
\item{ps_plot_dir}{directory with plot files}
}
\value{
vec_plot_result vector with plot filenames
}
\description{
Given a directory \code{ps_plot_dir} with pdf and png plot files, the list
of filenames is returned. If the directory contains a file with a name matching
the regular expression "^ge_plot_report.*\\.pdf$", this file is excluded.
}
