% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{get_qgert_logger}
\alias{get_qgert_logger}
\title{Create log4r Logger for package}
\usage{
get_qgert_logger(ps_logfile = "qgert.log", ps_level = "FATAL")
}
\arguments{
\item{ps_logfile}{name of the logfile}

\item{ps_level}{logger level}
}
\value{
qgert_logger
}
\description{
Create log4r Logger for package
}
\examples{
qgert_logger <- get_qgert_logger()
}
