% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toplist_bull_bvch.R
\name{get_vec_resultcols_bvch}
\alias{get_vec_resultcols_bvch}
\title{Default Result Columns}
\usage{
get_vec_resultcols_bvch()
}
\value{
The columns in the result file
}
\description{
The default for the columns shown in the result xlsx file.
}
