% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{qgert_log_debug}
\alias{qgert_log_debug}
\title{Wrapper for log4r debug}
\usage{
qgert_log_debug(plogger, ps_caller, ps_msg)
}
\arguments{
\item{plogger}{log4r logger object}

\item{ps_msg}{logging message}
}
\description{
Wrapper for log4r debug
}
\examples{
\dontrun{
qgert_logger <- get_qgert_logger()
qgert_log_level(qgert_logger, 'DEBUG')
qgert_log_debug(qgert_logger, 'Examples', 'test message')
}
}
