% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{qgert_log_error}
\alias{qgert_log_error}
\title{Wrapper for log4r error}
\usage{
qgert_log_error(plogger, ps_caller, ps_msg)
}
\arguments{
\item{plogger}{log4r logger object}

\item{ps_caller}{function from which we are called}

\item{ps_msg}{logging message}
}
\description{
Wrapper for log4r error
}
\examples{
\dontrun{
qgert_logger <- get_qgert_logger()
qgert_log_level(qgert_logger, 'ERROR')
qgert_log_error(qgert_logger, 'Examples', 'test message')
}
}
