% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{qgert_log_level}
\alias{qgert_log_level}
\title{Wrapper to set the level of a logger}
\usage{
qgert_log_level(plogger, ps_level = c("DEBUG", "INFO", "WARN", "ERROR",
  "FATAL"))
}
\arguments{
\item{plogger}{log4r logger object}

\item{ps_level}{new level of plogger}
}
\description{
Wrapper to set the level of a logger
}
\examples{
\dontrun{
qgert_logger <- get_qgert_logger()
qgert_log_level(qgert_logger, 'INFO')
}
}
