% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{qgert_log_warn}
\alias{qgert_log_warn}
\title{Wrapper for log4r warn}
\usage{
qgert_log_warn(plogger, ps_caller, ps_msg)
}
\arguments{
\item{plogger}{log4r logger object}

\item{ps_msg}{logging message}
}
\description{
Wrapper for log4r warn
}
\examples{
\dontrun{
qgert_logger <- get_qgert_logger()
qgert_log_level(qgert_logger, 'WARN')
qgert_log_warn(qgert_logger, 'Examples', 'test message')
}
}
