% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toplist_bull_bvch.R
\name{read_top_list_info3}
\alias{read_top_list_info3}
\title{Read Toplist Input From CSV-Files}
\usage{
read_top_list_info3(
  pl_breed_input,
  ptbl_trait,
  pvec_resultcols = get_vec_resultcols_bvch(),
  pb_debug = FALSE,
  plogger = NULL
)
}
\arguments{
\item{pl_breed_input}{list with input information about breeds}

\item{ptbl_trait}{tibble with trait information}

\item{pb_debug}{flag for debugging mode}

\item{plogger}{logger object}
}
\value{
list of toplist input
}
\description{
The content of the csv-files are read into a tibble and all tibbles are
combined into a list.
}
\details{
--- Reading TopList Input ------------------------------------------------- #


This function assumes that there is a separate csv-file for each breed.
}
