% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sort_rt_list.R
\name{split_gsruns_sorted_rt}
\alias{split_gsruns_sorted_rt}
\title{Split GsRuns According to Sorted Runtimes}
\usage{
split_gsruns_sorted_rt(
  ps_rt_in_file,
  ps_out_dir,
  pn_nr_split,
  ps_out_file_stem = "gsSortedRuns.txt"
)
}
\arguments{
\item{ps_rt_in_file}{gs-job runtime input file}

\item{ps_out_dir}{output directory where the splitted jobs should be written to}

\item{pn_nr_split}{number of splitted job files.}

\item{ps_out_file_stem}{stem of output filename}
}
\description{
Given an input file of runtimes per gs-jobs, the list of gs-jobs
will be sorted and output in a given number of splits of the
complete number of gs-jobs.
}
\details{
---
title: Sort a List of gsRuns According To Their Runtime
date:  "2019-10-30"
---
}
\examples{
\dontrun{
split_gsruns_sorted_rt(ps_rt_in_file = "work/rank_gsRuns_runtime.out", ps_out_dir = "work", pn_nr_split = 10)
}
}
