% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toplist_bull_bvch.R
\name{write_tl_xlsx}
\alias{write_tl_xlsx}
\title{Write Toplist Information to xlsx}
\usage{
write_tl_xlsx(pl_toplist, ptbl_trait, ps_xlsx_file = "example.xlsx")
}
\arguments{
\item{pl_toplist}{list with top list information}

\item{ptbl_trait}{tibble with trait name}

\item{ps_xlsx_file}{name of result file}
}
\description{
The extracted toplist information is written to an xlsx file.
}
\details{
--- Writing TopList To xlsx ----------------------------------------------- #
}
