% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_plot_report_generic.R
\name{create_ge_plot_report}
\alias{create_ge_plot_report}
\title{Create Report With Plots From Two GE Periods}
\usage{
create_ge_plot_report(ps_gedir, ps_archdir, ps_trgdir, ps_templ,
  ps_report_text, ps_rmd_report = "ge_plot_report.Rmd",
  pb_keep_src = FALSE, pb_session_info = TRUE, pb_debug = FALSE,
  plogger = NULL)
}
\arguments{
\item{ps_gedir}{directory with plots of current GE round}

\item{ps_archdir}{archive directory with plots from previous GE}

\item{ps_trgdir}{target directory where plot files from archive are extracted, relative to ps_gedir}

\item{ps_templ}{path to Rmarkdown template file}

\item{ps_report_text}{text that is included in the report before plotting}

\item{ps_rmd_report}{name of report source file}

\item{pb_keep_src}{should Rmd source be kept}

\item{pb_session_info}{should session_info be included}

\item{pb_debug}{flag indicating whether debug info is printed}

\item{plogger}{log4r logger object}
}
\description{
Based on a Rmarkdown Template (ps_templ) document all plots
in given GE directory (ps_gedir) are taken. For a given plot
in directory ps_gedir, a corresponding plot with the same
filename is searched in an archive directory (ps_archdir).
If such a plot is found, the two corresponding plots are
shown side-by-side in the generated Rmarkdown report.
}
\examples{
\dontrun{
create_ge_plot_report(ps_gedir     = "{path to comparison dir}",
                    ps_archdir     = "{path to archive dir}",
                    ps_trgdir      = "1904/compareBull",
                    ps_templ       = "inst/templates/compare_plots.Rmd.template",
                    ps_report_text = "## Comparison Of Plots\\nPlots compare estimates ...",
                    ps_rmd_report  = 'ge_plot_report_fbk_compareBulls_bv.Rmd',
                    pb_debug       = TRUE)

}
}
