% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toplist_bull_bvch.R
\name{get_vec_non_traits}
\alias{get_vec_non_traits}
\title{Vector of Non-Trait Columns}
\usage{
get_vec_non_traits()
}
\value{
Vector of column names that are not to be treated as traits
}
\description{
This function returns a fixed vector of columns that are not treated as
traits.
}
\details{
--- Defaults and Constants ------------------------------------------------- #
}
