% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spin_sh.R
\name{spin_sh}
\alias{spin_sh}
\title{Spin a bash script into Rmd}
\usage{
spin_sh(ps_sh_hair,
  ps_out_rmd = paste0(basename(fs::path_ext_remove(ps_sh_hair)), ".Rmd"),
  pb_knit = TRUE, pobi_output_format = NULL, pb_keep_rmd = FALSE)
}
\arguments{
\item{ps_sh_hair}{path to the bash scirpt}

\item{ps_out_rmd}{output Rmd file}

\item{pb_knit}{flag whether generated Rmd should be rendered}

\item{pobi_output_format}{desired output format passed to rmarkdown::render()}

\item{pb_keep_rmd}{should Rmd source file be kept}
}
\description{
This function is derived from \code{knitr::spin()} for R-scripts and
is adapted to do a similar conversion of bash scripts into Rmd files.
The conversion is done in a simple way and almost no features of the
original are provided. This is just a very crude way to have a look
at documentation of a shell script in a nicer format such as html.
When specified, the generated Rmd is rendered to html using
\code{rmarkdown::render()}.
}
\examples{
s_test_script <- system.file('extdata', 'test_script.sh', package = 'qgert')
s_test_out <- file.path(tempdir(), 'test_script.Rmd')
spin_sh(ps_sh_hair = s_test_script, ps_out_rmd = s_test_out)
unlink(s_test_out)

}
